Code organization by creating "utils.py" and "core.py" files. Some errors fixed on README examples. Using find_packages in "setup.py" file.
